+++ internal/acceptance_tests/blocks/logging_papertrail_nested_compute_format.tf
@@ -1,0 +1,6 @@
+logging_papertrail {
+  name    = "{{.LOGGING_PAPERTRAIL_NAME}}"
+  address = "logs.papertrailapp.com"
+  port    = 12345
+  format  = "%h %l %u %t \"%r\" %>s %b"
+}